+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 Changelog Application Jobsearch
 ===============================
+## 1.5.0
+* Search-contextual switch for typeahead
+* Fixes an issue with taxonomy search
+
 ## 1.4.3
-Buggfixes for freetext search and location.
+* Buggfixes for freetext search and location.
 
 ## 1.4.2
 * Buggfixes and documentation updates